pyproject.toml
[project]
name = "grounded-rag-evaluator"
version = "0.1.0"
description = "Production-grade RAG system with citation-backed answers and evaluation harness"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "streamlit>=1.56.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
qa = "scripts.qa:main"

